🚚 급식 조회 경로 변경
