--- src/main/kotlin/team/gsm/flooding/domain/lunch/controller/dto/response/FetchLunchResponse.kt
@@ -1,5 +1,0 @@
-package team.gsm.flooding.domain.lunch.controller.dto.response
-
-data class FetchLunchResponse(
-	val menu: List<String>,
-)
