+++ src/main/kotlin/team/gsm/flooding/domain/neis/dto/response/FetchMealInfoResponse.kt
@@ -1,0 +1,5 @@
+package team.gsm.flooding.domain.neis.dto.response
+
+data class FetchLunchResponse(
+	val menu: List<String>,
+)
